[fuzz] split out compress_fuzzer into 5 smaller tests
ninja test shows the following total running times:

2/7 Test #2: compress_fuzzer ..................   Passed    0.67 sec
3/7 Test #3: example_small_fuzzer .............   Passed    0.55 sec
4/7 Test #4: example_large_fuzzer .............   Passed    0.12 sec
5/7 Test #5: example_flush_fuzzer .............   Passed    0.20 sec
6/7 Test #6: example_dict_fuzzer ..............   Passed    1.16 sec

The patch also restricts all the fuzzers to process less than 1Mb input.
We will probably need to restrict the dict_fuzzer to less than 1Mb.
